Fix large sanitizer findings
Address the remaining findings from the larger sanitizer matrix: harden
integer-domain width arithmetic, avoid invalid marked-link casts, protect RBS
stop state, and make thread-runner cleanup tolerate null terminators.

+++ gecode/int/var-imp/delta.hpp
@@ -52,7 +52,7 @@ namespace Gecode { namespace Int {
   }
   forceinline unsigned int
   IntDelta::width(void) const {
-    return static_cast<unsigned int>(_max - _min + 1);
+    return static_cast<unsigned int>(_max) - static_cast<unsigned int>(_min) + 1U;
   }
   forceinline bool
   IntDelta::any(void) const {
